Fix falsey API longitutde/latitude bug
Previously, the API would have thrown an error handling user locations
with longitude zero or latitude zero (because these values are
considered false by js). This commit adds a test to make sure that
only nonexistent longitude or latitude throws an error and that zero
longitude and latitude are accepted.